projet_biometrique/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
pyserial==3.5
requests==2.31.0
